# docs/api/swagger.yaml
basePath: /api/v1
...
host: localhost:8080
info:
  contact:
    email: hello@goatflow.io
    name: GoatFlow Support
    url: https://goatflow.io/support
  description: GoatFlow Ticket System REST API
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://goatflow.io/terms/
  title: GoatFlow API
  version: "1.0"
paths:
...
  /tickets:
    get:
      consumes:
      - application/json
      description: Retrieve a paginated list of tickets with optional filtering, sorting,
        and search
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page (max 100)
        in: query
        name: per_page
        type: integer
      - default: 20
        description: Alias for per_page (max 100)
        in: query
        name: limit
        type: integer
      - description: Offset for pagination (alternative to page)
        in: query
        name: offset
        type: integer
      - description: Filter by status
        enum:
        - open
        - closed
        - pending
        in: query
        name: status
        type: string
      - description: Filter by queue ID
        in: query
        name: queue_id
        type: integer
      - description: Filter by priority ID
        in: query
        name: priority_id
        type: integer
      - description: Filter by customer user ID/login
        in: query
        name: customer_user_id
        type: string
      - description: Filter by assigned agent user ID
        in: query
        name: assigned_user_id
        type: integer
      - description: Search in ticket number, title, customer
        in: query
        name: search
        type: string
      - default: created
        description: Sort field
        enum:
        - created
        - updated
        - priority
        - tn
        in: query
        name: sort
        type: string
      - default: desc
        description: Sort order
        enum:
        - asc
        - desc
        in: query
        name: order
        type: string
      - description: 'Include related data (comma-separated: article_count, last_article)'
        in: query
        name: include
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of tickets with pagination
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List tickets
      tags:
      - Tickets
    post:
      consumes:
      - application/json
      description: Create a new ticket with optional initial article
      parameters:
      - description: Ticket data (title, queue_id, priority_id, customer_user_id,
          article)
        in: body
        name: ticket
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created ticket
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create ticket
      tags:
      - Tickets
...
securityDefinitions:
  BearerAuth:
    description: API token (Bearer gf_xxx) or JWT token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"