# internal/api/ticket_list_handler.go
package api

import (
	"fmt"
	"log"
	"net/http"
	"strconv"
	"strings"

	"github.com/gin-gonic/gin"

	"github.com/goatkit/goatflow/internal/platform/database"
	"github.com/goatkit/goatflow/internal/platform/lookups"
)

// TicketListResponse represents the response for ticket list API.
type TicketListResponse struct {
	Success    bool                     `json:"success"`
	Data       []map[string]interface{} `json:"data"`
	Pagination PaginationInfo           `json:"pagination"`
	Error      string                   `json:"error,omitempty"`
}

// PaginationInfo contains pagination metadata.
type PaginationInfo struct {
	Page       int  `json:"page"`
	PerPage    int  `json:"per_page"`
	Total      int  `json:"total"`
	TotalPages int  `json:"total_pages"`
	HasNext    bool `json:"has_next"`
	HasPrev    bool `json:"has_prev"`
}

// HandleListTicketsAPI handles GET /api/v1/tickets.
//
//	@Summary		List tickets
//	@Description	Retrieve a paginated list of tickets with optional filtering, sorting, and search
//	@Tags			Tickets
//	@Accept			json
//	@Produce		json
//	@Param			page				query		int		false	"Page number"						default(1)
//	@Param			per_page			query		int		false	"Items per page (max 100)"			default(20)
//	@Param			limit				query		int		false	"Alias for per_page (max 100)"		default(20)
//	@Param			offset				query		int		false	"Offset for pagination (alternative to page)"
//	@Param			status				query		string	false	"Filter by status"					Enums(open, closed, pending)
//	@Param			queue_id			query		int		false	"Filter by queue ID"
//	@Param			priority_id			query		int		false	"Filter by priority ID"
//	@Param			customer_user_id	query		string	false	"Filter by customer user ID/login"
//	@Param			assigned_user_id	query		int		false	"Filter by assigned agent user ID"
//	@Param			search				query		string	false	"Search in ticket number, title, customer"
//	@Param			sort				query		string	false	"Sort field"						Enums(created, updated, priority, tn)	default(created)
//	@Param			order				query		string	false	"Sort order"						Enums(asc, desc)						default(desc)
//	@Param			include				query		string	false	"Include related data (comma-separated: article_count, last_article)"
//	@Success		200					{object}	map[string]interface{}	"List of tickets with pagination"
//	@Failure		401					{object}	map[string]interface{}	"Unauthorized"
//	@Failure		500					{object}	map[string]interface{}	"Internal server error"
//	@Security		BearerAuth
//	@Router			/tickets [get]
func HandleListTicketsAPI(c *gin.Context) {
	_, exists := c.Get("user_id")
	if !exists {
		if _, authExists := c.Get("is_authenticated"); !authExists {
			c.JSON(http.StatusUnauthorized, gin.H{
				"success": false,
				"error":   "Authentication required",
			})
			return
		}
	}

	// Parse pagination parameters
	page := 1
	perPage := 20

	if pageStr := c.Query("page"); pageStr != "" {
		if p, err := strconv.Atoi(pageStr); err == nil && p > 0 {
			page = p
		}
	}

	if perPageStr := c.Query("per_page"); perPageStr != "" {
		if pp, err := strconv.Atoi(perPageStr); err == nil && pp > 0 {
			perPage = pp
			// Cap at 100 items per page
			if perPage > 100 {
				perPage = 100
			}
		}
	}

	// Support 'limit' as alias for 'per_page'
	if limitStr := c.Query("limit"); limitStr != "" {
		if lim, err := strconv.Atoi(limitStr); err == nil && lim > 0 {
			perPage = lim
			if perPage > 100 {
				perPage = 100
			}
		}
	}

	// Support 'offset' - converts to page number
	if offsetStr := c.Query("offset"); offsetStr != "" {
		if off, err := strconv.Atoi(offsetStr); err == nil && off >= 0 {
			// Convert offset to page: page = (offset / perPage) + 1
			page = (off / perPage) + 1
		}
	}

	// Parse filter parameters
	filters := make(map[string]interface{})

	if status := c.Query("status"); status != "" {
		// Map status categories to state-id subqueries keyed by state type name
		switch status {
		case "open":
			filters["state_ids_sql"] = lookups.NewOpenStateIDsSQL
		case "closed":
			filters["state_ids_sql"] = lookups.ClosedStateIDsSQL
		case "pending":
			filters["state_ids_sql"] = lookups.PendingStateIDsSQL
		default:
			filters["state_name"] = status
		}
	}

	if queueID := c.Query("queue_id"); queueID != "" {
		if qid, err := strconv.Atoi(queueID); err == nil {
			filters["queue_id"] = qid
		}
	}

	if priorityID := c.Query("priority_id"); priorityID != "" {
		if pid, err := strconv.Atoi(priorityID); err == nil {
			filters["priority_id"] = pid
		}
	}

	if customerUserID := c.Query("customer_user_id"); customerUserID != "" {
		filters["customer_user_id"] = customerUserID
	}

	if assignedUserID := c.Query("assigned_user_id"); assignedUserID != "" {
		if auid, err := strconv.Atoi(assignedUserID); err == nil {
			filters["responsible_user_id"] = auid
		}
	}

	// Handle search parameter
	search := c.Query("search")

	// Parse sorting parameters
	sortField := c.DefaultQuery("sort", "created")
	sortOrder := c.DefaultQuery("order", "desc")

	// Map sort field names to database columns
	sortColumn := "t.create_time"
	switch sortField {
	case "created":
		sortColumn = "t.create_time"
	case "updated":
		sortColumn = "t.change_time"
	case "priority":
		sortColumn = "t.ticket_priority_id"
	case "tn":
		sortColumn = "t.tn"
	case "title":
		sortColumn = "t.title"
	}

	// Validate sort order
	if sortOrder != "asc" && sortOrder != "desc" {
		sortOrder = "desc"
	}

	// Parse include parameter for related data
	includes := strings.Split(c.Query("include"), ",")
	includeLastArticle := false
	includeArticleCount := false

	for _, inc := range includes {
		switch strings.TrimSpace(inc) {
		case "last_article":
			includeLastArticle = true
		case "article_count":
			includeArticleCount = true
		}
	}

	// Get database connection
	db, err := database.GetDB()
	if err != nil || db == nil {
		log.Printf("HandleListTicketsAPI: database unavailable: %v", err)
		c.JSON(http.StatusInternalServerError, gin.H{
			"success": false,
			"error":   "Database unavailable",
		})
		return
	}

	// Customers see only their own tickets; agents only tickets in queues they
	// can read. An empty access set matches nothing.
	scope, ok := resolveTicketReadScope(c, db, true)
	if !ok {
		return
	}

	// Build the query
	offset := (page - 1) * perPage

	// Base query
	query := `
		SELECT
			t.id,
			t.tn,
			t.title,
			t.queue_id,
			q.name as queue_name,
			t.ticket_state_id as state_id,
			ts.name as state_name,
			t.ticket_priority_id as priority_id,
			tp.name as priority_name,
			t.customer_user_id,
			t.customer_id,
			t.user_id,
			t.responsible_user_id,
			t.create_time as created_at,
			t.change_time as updated_at
		FROM ticket t
		LEFT JOIN queue q ON t.queue_id = q.id
		LEFT JOIN ticket_state ts ON t.ticket_state_id = ts.id
		LEFT JOIN ticket_priority tp ON t.ticket_priority_id = tp.id
		WHERE 1=1`

	args := []interface{}{}

	// Add filters to query
	if stateIDsSQL, ok := filters["state_ids_sql"].(string); ok {
		query += " AND t.ticket_state_id IN (" + stateIDsSQL + ")"
	}

	if queueID, ok := filters["queue_id"].(int); ok {
		query += " AND t.queue_id = ?"
		args = append(args, queueID)
	}

	if priorityID, ok := filters["priority_id"].(int); ok {
		query += " AND t.ticket_priority_id = ?"
		args = append(args, priorityID)
	}

	if customerUserID, ok := filters["customer_user_id"].(string); ok && customerUserID != "" {
		query += " AND t.customer_user_id = ?"
		args = append(args, customerUserID)
	}

	if responsibleUserID, ok := filters["responsible_user_id"].(int); ok {
		query += " AND t.responsible_user_id = ?"
		args = append(args, responsibleUserID)
	}

	scopeCond, scopeArgs := scope.filter("t")
	query += " AND " + scopeCond
	args = append(args, scopeArgs...)

	// Add search condition
	if search != "" {
		query += " AND (LOWER(t.title) LIKE LOWER(?) OR t.tn = ?)"
		args = append(args, "%"+search+"%", search)
	}

	// Get total count
	countQuery := "SELECT COUNT(*) FROM (" + query + ") as count_query"
	var total int
	err = db.QueryRow(database.ConvertPlaceholders(countQuery), args...).Scan(&total)
	if err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{
			"success": false,
			"error":   "Failed to count tickets",
		})
		return
	}

	// Add sorting and pagination
	query += fmt.Sprintf(" ORDER BY %s %s", sortColumn, strings.ToUpper(sortOrder))
	query += " LIMIT ? OFFSET ?"
	args = append(args, perPage, offset)

	// Execute main query
	rows, err := db.Query(database.ConvertPlaceholders(query), args...)
	if err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{
			"success": false,
			"error":   "Failed to fetch tickets",
		})
		return
	}
	defer rows.Close()

	// Process results
	tickets := []map[string]interface{}{}
	for rows.Next() {
		var ticket struct {
			ID                int64   `json:"id"`
			TN                string  `json:"tn"`
			Title             string  `json:"title"`
			QueueID           int     `json:"queue_id"`
			QueueName         string  `json:"queue_name"`
			StateID           int     `json:"state_id"`
			StateName         string  `json:"state_name"`
			PriorityID        int     `json:"priority_id"`
			PriorityName      string  `json:"priority_name"`
			CustomerUserID    *string `json:"customer_user_id"`
			CustomerID        *string `json:"customer_id"`
			UserID            int     `json:"user_id"`
			ResponsibleUserID *int    `json:"responsible_user_id"`
			CreatedAt         string  `json:"created_at"`
			UpdatedAt         string  `json:"updated_at"`
		}

		err := rows.Scan(
			&ticket.ID,
			&ticket.TN,
			&ticket.Title,
			&ticket.QueueID,
			&ticket.QueueName,
			&ticket.StateID,
			&ticket.StateName,
			&ticket.PriorityID,
			&ticket.PriorityName,
			&ticket.CustomerUserID,
			&ticket.CustomerID,
			&ticket.UserID,
			&ticket.ResponsibleUserID,
			&ticket.CreatedAt,
			&ticket.UpdatedAt,
		)
		if err != nil {
			log.Printf("HandleListTicketsAPI: scan ticket row: %v", err)
			c.JSON(http.StatusInternalServerError, gin.H{
				"success": false,
				"error":   "Failed to fetch tickets",
			})
			return
		}

		// Convert to map for flexible response
		ticketMap := map[string]interface{}{
			"id":            ticket.ID,
			"tn":            ticket.TN,
			"ticket_number": ticket.TN,
			"title":         ticket.Title,
			"queue_id":      ticket.QueueID,
			"queue_name":    ticket.QueueName,
			"state_id":      ticket.StateID,
			"state_name":    ticket.StateName,
			"priority_id":   ticket.PriorityID,
			"priority_name": ticket.PriorityName,
			"created_at":    ticket.CreatedAt,
			"updated_at":    ticket.UpdatedAt,
			"create_time":   ticket.CreatedAt,
			"update_time":   ticket.UpdatedAt,
		}

		// Add optional fields
		if ticket.CustomerUserID != nil {
			ticketMap["customer_user_id"] = *ticket.CustomerUserID
		} else {
			ticketMap["customer_user_id"] = ""
		}

		if ticket.CustomerID != nil {
			ticketMap["customer_id"] = *ticket.CustomerID
		} else {
			ticketMap["customer_id"] = ""
		}

		ticketMap["user_id"] = ticket.UserID

		if ticket.ResponsibleUserID != nil {
			ticketMap["responsible_user_id"] = *ticket.ResponsibleUserID
		} else {
			ticketMap["responsible_user_id"] = nil
		}

		// Add included relations if requested
		if includeArticleCount {
			var count int
			countErr := db.QueryRow(database.ConvertPlaceholders(
				"SELECT COUNT(*) FROM article WHERE ticket_id = ?",
			), ticket.ID).Scan(&count)
			if countErr == nil {
				ticketMap["article_count"] = count
			} else {
				ticketMap["article_count"] = 0
			}
		}

		if includeLastArticle {
			// Get last article for this ticket
			var lastArticle struct {
				Subject   *string
				CreatedAt string
			}
			articleErr := db.QueryRow(database.ConvertPlaceholders(`
				SELECT adm.a_subject, a.create_time
				FROM article a
				LEFT JOIN article_data_mime adm ON a.id = adm.article_id
				WHERE a.ticket_id = ?
				ORDER BY a.create_time DESC
				LIMIT 1
			`), ticket.ID).Scan(&lastArticle.Subject, &lastArticle.CreatedAt)

			if articleErr == nil {
				subject := ""
				if lastArticle.Subject != nil {
					subject = *lastArticle.Subject
				}
				ticketMap["last_article"] = map[string]interface{}{
					"subject":    subject,
					"created_at": lastArticle.CreatedAt,
				}
			} else {
				ticketMap["last_article"] = nil
			}
		}

		tickets = append(tickets, ticketMap)
	}
	if err := rows.Err(); err != nil {
		log.Printf("HandleListTicketsAPI: iterate ticket rows: %v", err)
		c.JSON(http.StatusInternalServerError, gin.H{
			"success": false,
			"error":   "Failed to fetch tickets",
		})
		return
	}

	// Calculate pagination info
	totalPages := (total + perPage - 1) / perPage
	hasNext := page < totalPages
	hasPrev := page > 1

	// Return response
	c.JSON(http.StatusOK, TicketListResponse{
		Success: true,
		Data:    tickets,
		Pagination: PaginationInfo{
			Page:       page,
			PerPage:    perPage,
			Total:      total,
			TotalPages: totalPages,
			HasNext:    hasNext,
			HasPrev:    hasPrev,
		},
	})
}